Ignore SIGPIPE while running
We must be robust in the face of one of our reducer scripts dying while we're
still trying to communicate with it. We are set up to handle `Err` results
properly, but we can't tolerate receiving SIGPIPE and getting killed because we
attempted to write to the defunct child process's stdin.