Add a tag for every interesting test case
This prevents a race where a worker tries to pull from the supervisor's
repo to check out its reduction's seed's commit, but the supervisor has already
moved on to a new HEAD commit, and therefore the seed's commit is not pointed to
by any refspec and the worker won't get it in the fetch. We avoid this by always
making a tag for every interesting test case's commit.