add tests for and fix bug in regexp-based reducers
Before this change, all regexp-based reducers would first generate a
reduction by removing all of the matching lines from the seed file. They
would then proceed to generate reductions that removed all matching
lines except the first, and then the first and second, and so on. This
changes the implementation to only remove one matching line at a time,
which I believe was the original intention. (Note that it appears the
original creduce pass_blank.pm would remove all blank lines to generate
one reduction and then remove all lines that started with '#' to
generate a second reduction.)