fix bug where balanced reducers would read stdin too much
For each balanced pair, balanced reducers will skip the reduction consisting of
removing the characters between the pair's indices if the pair is of the form
(n, n+1). However, the initial implementation would have incorrectly already
read the next reduction output location from stdin, thus breaking the protocol.
This would result in hangs where the reducers would simply stop generating
reductions. The fix is simple: only read the output location if the reduction is
going to be generated.