Extend the Reducer trait to be cognizant of concurrency
This commit gives the Reducer trait the ability to reason about reductions that
are being tested concurrently and builds what was the `DontRepeatYourself`
combinator into the trait directly. Why build that into the trait directly?
First, so that more complex reducers can opt out of that behavior, if it isn't a
good heuristic for their reductions. Second, so that it can be implemented in
O(1) time.

This commit does not port reducer scripts over to this new style -- that
requires updating the IPC protocol.

Nor does this commit make the actors leverage
`Reducer::next_state_on_interesting` yet; that is left for a follow up.